Add guide on AWS authentication via EC2 IAM role
Since Architect doesn't support using an EC2 IAM role to
authenticate AWS API calls, and adding support would require
big changes to core functionality, adding a note to the
documentation can help the occasional dev who finds themself
needing using this form of authentication.

Since this is principally an example of how to handle a particular
form of authentication, I included it in the section on setting
up AWS credentials. I made the example as simple as I could,
and wrote it in JS, because that's the language I'm using, so
it's the one that I've actually tested and seen work.